requirements.txt
face_recognition==1.3.0
numpy==1.24.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
supabase==2.4.5
python-dotenv
redis-dict